app/masterapp/requirements.txt

Django==3.2.15
celery==4.4.0
psycopg2_binary==2.9.3
redis==3.4.1
djangorestframework==3.14.0
django-cors-headers==3.13.0
djangorestframework-api-key==2.2.0
